pyproject.toml
[project]
name = "youdub-webui"
version = "0.1.0"
description = "Translate and dub YouTube videos to Chinese using AI"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { text = "MIT" }
authors = [{ name = "liuzhao1225" }]
keywords = ["youtube", "translation", "dubbing", "ai", "tts", "whisper"]

dependencies = [
    # Web UI
    "gradio",

    # Logging
    "loguru>=0.7.0,<1.0.0",


    # Configuration
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",

    # Video downloading
    "yt-dlp[default,curl-cffi,secretstorage]",
    "curl_cffi",
    "yt-dlp-invidious",
    "bgutil-ytdlp-pot-provider",
    "google-genai>=1.50.0",

    # Audio processing
    "scipy>=1.11.0,<1.14.0",
    "librosa>=0.10.0,<0.11.0",
    "audiostretchy>=1.3.5",
    "numpy<2.0.0",

    # Speech recognition - WhisperX from PyPI
    "whisperx>=3.1.0",

    # Speaker diarization (required by bytedance TTS for speaker matching)
    # Note: Keep <4.0.0 for Torchaudio 2.1.x compatibility
    "pyannote.audio>=3.1.0,<4.0.0",

    # Translation API
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "transformers>=4.38.0,<4.40.0",

    # Image processing
    "Pillow>=10.0.0,<11.0.0",

    # HTTP requests (for bytedance TTS)
    "requests>=2.31.0,<3.0.0",

    # Bilibili upload
    # Bilibili upload
    "bilibili_toolman",

    # PyTorch ecosystem (CUDA 12.1)
    "torch==2.1.2",
    "torchaudio==2.1.2",
]

[project.optional-dependencies]
# Demucs for audio source separation (vocals/instruments)
# Note: Demucs has strict torchaudio<2.1 requirement that conflicts with whisperx
# To install demucs manually (it works fine with newer torchaudio in practice):
#   uv pip install --no-deps git+https://github.com/facebookresearch/demucs
#   uv pip install dora-search einops "julius>=0.2.3" "lameenc>=1.2" openunmix
demucs = ["dora-search", "einops", "julius>=0.2.3", "lameenc>=1.2", "openunmix"]

# XTTS for voice cloning TTS
# Note: Resolving TTS dependency automatically fails due to strict numpy==1.22.0 requirement.
# Install manually: uv pip install --no-deps "TTS>=0.22.0"
xtts = []

# Development dependencies
dev = ["pytest>=7.0.0", "black>=23.0.0", "ruff>=0.1.0"]

[project.scripts]
youdub = "youdub.app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["youdub*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
target-version = ["py310"]

# --- UV Package Manager Configuration ---

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "sys_platform == 'linux'" }]
torchaudio = [{ index = "pytorch-cu121", marker = "sys_platform == 'linux'" }]
